+++ packages/emoji/CHANGELOG.md
@@ -9,7 +9,7 @@
 
 #### 🚀 Updates
 
-- Updated to support Emoji v13, which includes 117 new emojis.
+- Updated to support Emoji v13.1, which includes 117+ new emojis.
 - Updated to support different shortcodes using pre-defined presets.
 - Added `shortcodes` option to `useEmojiData`.